+++ push_swap.h
@@ -44,6 +44,7 @@ void		rrr(t_element *a, t_element *b, int top_a, int top_b);
 //
 int checker(char **av, int ac,int **tab,int *len);
 int check_if_sorted(int *vars, int len);
+// void print(t_stack a,t_stack b,int len);
 //////swap5 and 3
 int		bigestone(t_element *a, int len);
 int		smalestone(t_element *a, int len);
@@ -54,7 +55,7 @@ void sort_big(t_stack *a,t_stack *b,int range);
 //swap more 10
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t count, size_t size);
-int	ft_atoi(const char *str);
+long	ft_atoi(const char *str);
 size_t	ft_strlen(const char *s);
 int	ft_strncmp(const char *str1, const char *str2, size_t n);
 char	**ft_split(char *s, char c);
